Move MockJoystick into source tree instead of test tree
